+++ bauble/src/context.rs
